# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Back End Coursework API
  description: A simple e-commerce API.
  contact:
    name: Joseph Lees
    email: leesj4@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Development server
...
paths:
...
  /products:
    summary: Represents an individual product
    description: >
      This resource represents an individual product in the system.
      Each product is identified by a numeric `id`.
    post:
      tags:
        - Products
      description:
        Create a new product.
      requestBody:
        description: new product information
        required: true
        content:
          application/json:
            schema:
              $ref: ./product.json#/definitions/product
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: ./product.json#/definitions/productCreated
        '403':
          description: Non-admin users can't create products.
        '500':
          description: Internal server error.
    get:
      tags:
        - Products
      description:
        Get all products.
      responses:
        '200':
          description: View of all products records
        '404':
          description: No products found in the DB.
        '500':
          description: Internal server error.
  /products/{id}:
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the product
    get:
      tags:
        - Products
      description:
        Get product by id.
      responses:
        '200':
          description: View of a products record.
        '404':
          description: Product was not found in the DB.
        '500':
          description: Internal server error.
    put:
      tags:
        - Products
      description:
        Update product by id.
      requestBody:
        description: updated product information
        required: true
        content:
          application/json:
            schema:
              $ref: ./product.json#/definitions/productUpdate
      responses:
        '200':
          description: Product was updated successfully.
          content:
            application/json:
              schema:
                $ref: ./product.json#/definitions/productUpdated
        '403':
          description: Non-admin users can't update products.
        '404':
          description: Product was not found in the DB.
        '500':
          description: Internal server error.
    delete:
      tags:
        - Products
      description:
        Delete product by id.
      responses:
        '204':
          description: Product was deleted successfully.
        '403':
          description: Non-admin users can't delete products.
        '404':
          description: Product was not found in the DB.
        '500':
          description: Internal server error.
...
  /categories/{id}:
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the category
    get:
      tags:
        - Categories
      description:
        Get category by id.
      responses:
        '200':
          description: View of a category record.
        '403':
          description: Non-admins cannot view categories.
        '404':
          description: Category was not found in the DB.
        '500':
          description: Internal server error.
    put:
      tags:
        - Categories
      description:
        Update category by id.
      requestBody:
        description: updated category information
        required: true
        content:
          application/json:
            schema:
              $ref: ./category.json#/definitions/categoryUpdate
      responses:
        '200':
          description: Category was updated successfully.
          content:
            application/json:
              schema:
                $ref: ./category.json#/definitions/categoryUpdated
        '403':
          description: Non-admin users can't update categories.
        '404':
          description: Category was not found in the DB.
        '500':
          description: Internal server error.
    delete:
      tags:
        - Categories
      description:
        Delete category by id.
      responses:
        '204':
          description: Category was deleted successfully.
        '403':
          description: Non-admin users can't delete categories.
        '404':
          description: Category was not found in the DB.
        '500':
          description: Internal server error.
  /orders:
    summary: Represents a single order
    description: >
      This resource represents an order in the system.
      Each order is identified by a numeric `id`.
    post:
      tags:
        - Orders
      description:
        Create a new order.
      requestBody:
        description: new order information
        required: true
        content:
          application/json:
            schema:
              $ref: ./order.json#/definitions/order
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: ./order.json#/definitions/orderCreated
        '400':
          description: The product has already sold.
        '403':
          description: Both users and admins can create orders, kept permission check incase of future change to roles.
        '404':
          description: Product or address was not found in the DB.
        '500':
          description: Internal server error.
    get:
      tags:
        - Orders
      description:
        Get all orders, or user orders if user is not an admin.
      responses:
        '200':
          description: View of order records
        '404':
          description: No orders found in the DB.
        '500':
          description: Internal server error.
  /orders/{id}:
    parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric id of the order
    get:
      tags:
        - Orders
      description:
        Get order by id.
      responses:
        '200':
          description: View of an order record.
        '403':
          description: Non-admin users can only view their own orders.
        '404':
          description: Order was not found in the DB.
        '500':
          description: Internal server error.
    put:
      tags:
        - Orders
      description:
        Update order by id.
      requestBody:
        description: updated order information
        required: true
        content:
          application/json:
            schema:
              $ref: ./order.json#/definitions/orderUpdate
      responses:
        '200':
          description: Order was updated successfully.
          content:
            application/json:
              schema:
                $ref: ./order.json#/definitions/orderUpdated
        '403':
          description: Non-admin users can't update orders.
        '404':
          description: Order was not found in the DB.
        '500':
          description: Internal server error.
    delete:
      tags:
        - Orders
      description:
        Delete order by id.
      responses:
        '204':
          description: Order was deleted successfully.
        '403':
          description: Non-admin users can't delete orders.
        '404':
          description: Order was not found in the DB.
        '500':
          description: Internal server error.